Async Payjoin send and receives
The whole point of Payoin V2 is to save the state of a payjoin and resume it
whenever you have internet. Our first implementation doesn't do that since the
logic is limited to the Send or Receive UI context where it's created and it
disappears when the app goes out of memory.

This change persists that information in HiveStorage, pauses sessions on app
background, and resumes sessions on app foregroudn, allowing async payjoin to
happen.

There was a bug where Sender's didn't persist information sent in their first
request, so once they were persisted and resumed they were out of sync with
receivers. This has been fixed in an upcoming release of payjoin-flutter
and so requires that update in order to function.

In order to test, start a receiver, background that app, then start a sender
to that receiver, then background that app, then resume the receiver, then stop it,
then resume the sender. The payjoin should flow good as ever as long as there's
enough time to send the first payjoin message before stopping at each stage.