Don't display txid in toast
While we should create a success screen to display the Payjoin Success
with a link to mempool.space, I'm not comfortable linking directly from the
alert, since some people might not want to reveal their IP addres to
the explorer, and making a stopgap payjoin success screen seems like
an unnecessary hack when we plan to follow up with a proper payjoin history.

So this just gets rid of the txid that renders for now.