Enum unrecoverable Sender errors
Mark unrecoverable sender errors in storage so that they no longer spawn
isolates.

Spawning isolates and failing them immediately uses more resources than
necessary. This is therefore a performance improvement for the stability
of the app.

In the medium term we should switch on errors from payjoin-flutter, but
those errors are not yet available to switch on. This will reduce
the amount of error definition and handling we have to manually implement,
but since this is a potential immediate performance issue I'm addressing
it straight away.