Handle receiver unrecoverable errors
Receivers need to handle both switch errors or isolate errors. This change
takes the error handling out of the spawn try-catch, which would never receive
them and moves it into places where those errors would propagate.

Note that some of the switch errors like checkIsOwned might crop up
because of a wallet being improperly loaded, but for the time being we don't
have the granularity to distinguish between them.